Apply layout on TrainerPage
